kairos-take-home-0/requirements.txt
fastapi
uvicorn
requests
PyPDF2
python-dotenv
openai
anthropic
google-generativeai